[equ, gippo] обмен подарочными сертификатами. Изменился equ-api! (refs #4016)

+++ equ-clt/src/main/java/equ/clt/handler/atol/AtolHandler.java
@@ -551,7 +551,7 @@ public class AtolHandler extends CashRegisterHandler<AtolSalesBatch> {
                             if (dateReceipt == null || startDate == null || dateReceipt.compareTo(startDate) >= 0)
                                 currentSalesInfoList.add(new SalesInfo(false, nppGroupMachinery, numberCashRegister, numberZReport,
                                         dateReceipt, timeReceipt, numberReceipt, dateReceipt, timeReceipt, null, null, null, null,
-                                        null, null, barcodeItem, null, itemObject, null, quantityReceiptDetail, priceReceiptDetail,
+                                        null, (BigDecimal) null, barcodeItem, null, itemObject, null, quantityReceiptDetail, priceReceiptDetail,
                                         sumReceiptDetail, discountSumReceiptDetail, null, null, numberReceiptDetail, file.getName(), null));
                         } else if (isCancelDocument) {
                             cancelReceiptSet.add(numberReceipt);
