+++ equ-clt/src/main/java/equ/clt/handler/kristal/KristalHandler.java
@@ -1053,7 +1053,7 @@ public class KristalHandler extends CashRegisterHandler<KristalSalesBatch> {
 
                                                 if (dateReceipt == null || startDate == null || dateReceipt.compareTo(startDate) >= 0)
                                                     currentSalesInfoList.add(new SalesInfo(false, getNppGroupMachinery(directoryCashRegisterMap, directory, numberCashRegister), numberCashRegister,
-                                                            numberZReport, dateReceipt, timeReceipt, numberReceipt, dateReceipt, timeReceipt, null, null, null, sumCard, sumCash, null, barcode,
+                                                            numberZReport, dateReceipt, timeReceipt, numberReceipt, dateReceipt, timeReceipt, null, null, null, sumCard, sumCash, (BigDecimal) null, barcode,
                                                             idItem, null, null, quantity, price, sumReceiptDetail, null, discountSumReceipt, null, numberReceiptDetail, fileName, null));
                                             }
 
@@ -1148,7 +1148,7 @@ public class KristalHandler extends CashRegisterHandler<KristalSalesBatch> {
                                                     if (dateReceipt == null || startDate == null || dateReceipt.compareTo(startDate) >= 0)
                                                         currentSalesInfoList.add(new SalesInfo(false, getNppGroupMachinery(directoryCashRegisterMap, directory, numberCashRegister),
                                                                 numberCashRegister, numberZReport, dateReceipt, timeReceipt, numberReceipt, dateReceipt, timeReceipt, idEmployee,
-                                                                null, null, sumCard, sumCash, null, barcode, idItem, null, null, quantity, price, sumReceiptDetail, discountSumReceiptDetail,
+                                                                null, null, sumCard, sumCash, (BigDecimal) null, barcode, idItem, null, null, quantity, price, sumReceiptDetail, discountSumReceiptDetail,
                                                                 null, discountCard, numberReceiptDetail, fileName, null));
                                                 }
